Time: 29 ms (6.6%), Space: 55.8 MB (5.2%) - LeetHub

+++ 1976-number-of-ways-to-arrive-at-destination/1976-number-of-ways-to-arrive-at-destination.java
@@ -43,7 +43,7 @@ class Solution {
             int currNode = inner.node;
 
             for(List<Integer> list : adjancyList.get(currNode)) {
-                int innerDist = list.get(1);
+                long innerDist = list.get(1);
                 int innerNode = list.get(0);
                 if((currDist + innerDist) < distance[innerNode]) {
                     path[innerNode] = path[currNode];
